Fixes linter and renames files with _*
